requirements.txt
google-cloud-aiplatform
pandas
numpy
scikit-learn
joblib
dvc==3.63.0
dvc-gs==3.0.2
pytest
google-cloud-storage
gcsgs